coded refactor - moved array impl, clippy fix
