+++ elrond-codec/src/lib.rs
@@ -10,6 +10,7 @@ pub use alloc::vec::Vec;
 
 mod codec_err;
 mod default_traits;
+mod impl_array;
 mod nested_de;
 mod nested_de_input;
 mod nested_ser;
